Getting Yup initial values form QuestionaryComponentDefinition
