+++ src/components/questionary/questionaryComponents/Boolean/BooleanDefinition.tsx
@@ -20,4 +20,5 @@ export const booleanDefinition: QuestionaryComponentDefinition = {
   icon: <CheckBoxOutlineBlankIcon />,
   answerRenderer: ({ answer }) => <span>{answer.value}</span>,
   createYupValidationSchema: createBooleanValidationSchema,
+  getYupInitialValue: answer => answer.value || false,
 };
