+++ src/components/questionary/questionaryComponents/DatePicker/DatePickerDefinition.tsx
@@ -20,4 +20,5 @@ export const dateDefinition: QuestionaryComponentDefinition = {
   icon: <CheckBoxOutlineBlankIcon />,
   answerRenderer: ({ answer }) => <span>{answer.value}</span>,
   createYupValidationSchema: createDateValidationSchema,
+  getYupInitialValue: answer => answer.value || '',
 };
