+++ src/components/questionary/questionaryComponents/TextInput/TextInputDefinition.tsx
@@ -20,4 +20,5 @@ export const textInputDefinition: QuestionaryComponentDefinition = {
   icon: <ShortTextIcon />,
   answerRenderer: ({ answer }) => <span>{answer.value}</span>,
   createYupValidationSchema: createTextInputValidationSchema,
+  getYupInitialValue: answer => answer.value || '',
 };
